LayoutTests/http/tests/security/video-poster-cross-origin-crash-expected.txt

CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
>>>
